refactor[#345] : 주소 정보 api 요청 함수에 apiClient 적용 및 파일 위치 이동
