+++ src/entities/auth/constants/index.ts
@@ -5,3 +5,9 @@ export const SOCIAL_TYPE = {
   NAVER: "네이버",
   GOOGLE: "구글",
 };
+
+export const ADDRESSES_END_POINT = {
+  CURRENT_POSITION: ({ lat, lng }: { lat: number; lng: number }) =>
+    `/addresses/search-by-location?lat=${lat}&lng=${lng}`,
+  ADDRESS: (keyword: string) => `/addresses?keyword=${keyword}`,
+};
